added more exploration of api
Added pagination support and exploration of visitor endpoint. Decided to run with media and events initially when moving forward to set up pipeline. Visitor info can be added later if necessary for a visitor_dim table.

Commit reflects two steps in exploration, shown in two .md documents. An initial one and an updated one with the first visitor script adjusted and the second one dropped.